# backend_openapi_final.json
{"openapi":"3.1.0","info":{"title":"Todo App API","description":"Phase II Todo Full-Stack Web Application API","version":"1.0.0"},"paths":{"/health":{"get":{"tags":["Health"],"summary":"Health Check","description":"Health check endpoint to verify API is running.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"tags":["Root"],"summary":"Root","description":"Root endpoint with API information.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["Authentication"],"summary":"Signup","description":"Create a new user account.\n\n- **email**: Valid email address (RFC 5322)\n- **password**: Minimum 8 characters\n\nReturns user data and JWT token.","operationId":"signup_api_v1_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/signin":{"post":{"tags":["Authentication"],"summary":"Signin","description":"Sign in an existing user.\n\n- **email**: User's email address\n- **password**: User's password\n\nReturns user data and JWT token.","operationId":"signin_api_v1_auth_signin_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["Authentication"],"summary":"Refresh Token","description":"Refresh JWT token.\n\nRequires valid (but possibly expiring) JWT token in Authorization header.\nReturns new JWT token with extended expiry.","operationId":"refresh_token_api_v1_auth_refresh_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Current User","description":"Get current user profile.\n\nRequires valid JWT token in Authorization header.\nReturns authenticated user's profile information.","operationId":"get_current_user_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/tasks":{"get":{"tags":["Tasks","Tasks"],"summary":"List Tasks","description":"Get all tasks for authenticated user.\n\nArgs:\n    completed: Optional filter by completion status\n    offset: Pagination offset\n    limit: Number of records to return (max 100)\n    user_id: User ID from JWT token (injected by dependency)\n    session: Database session\n\nReturns:\n    List of tasks with count\n\nRaises:\n    401: If JWT token is invalid or missing","operationId":"list_tasks_api_v1_tasks_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"completed","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"description":"Filter by completion status","title":"Completed"},"description":"Filter by completion status"},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","maximum":10000,"minimum":0,"description":"Offset for pagination (max 10000)","default":0,"title":"Offset"},"description":"Offset for pagination (max 10000)"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Limit for pagination (max 100)","default":50,"title":"Limit"},"description":"Limit for pagination (max 100)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Tasks","Tasks"],"summary":"Create Task","description":"Create a new task for authenticated user.\n\nArgs:\n    request: Task creation request\n    user_id: User ID from JWT token (injected by dependency)\n    session: Database session\n\nReturns:\n    Created task\n\nRaises:\n    400: If validation fails (empty title, too long)\n    401: If JWT token is invalid or missing","operationId":"create_task_api_v1_tasks_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks/{task_id}":{"get":{"tags":["Tasks","Tasks"],"summary":"Get Task","description":"Get a specific task by ID.\n\nArgs:\n    task_id: Task UUID\n    user_id: User ID from JWT token (injected by dependency)\n    session: Database session\n\nReturns:\n    Task details\n\nRaises:\n    401: If JWT token is invalid or missing\n    404: If task not found or doesn't belong to user","operationId":"get_task_api_v1_tasks__task_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Tasks","Tasks"],"summary":"Update Task","description":"Update task title and/or description.\n\nArgs:\n    task_id: Task UUID\n    request: Task update request\n    user_id: User ID from JWT token (injected by dependency)\n    session: Database session\n\nReturns:\n    Updated task\n\nRaises:\n    400: If validation fails\n    401: If JWT token is invalid or missing\n    404: If task not found or doesn't belong to user","operationId":"update_task_api_v1_tasks__task_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Tasks","Tasks"],"summary":"Delete Task","description":"Delete a task permanently.\n\nArgs:\n    task_id: Task UUID\n    user_id: User ID from JWT token (injected by dependency)\n    session: Database session\n\nReturns:\n    No content (204)\n\nRaises:\n    401: If JWT token is invalid or missing\n    404: If task not found or doesn't belong to user","operationId":"delete_task_api_v1_tasks__task_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks/{task_id}/complete":{"patch":{"tags":["Tasks","Tasks"],"summary":"Toggle Task Completion","description":"Toggle task completion status.\n\nArgs:\n    task_id: Task UUID\n    request: Completion toggle request\n    user_id: User ID from JWT token (injected by dependency)\n    session: Database session\n\nReturns:\n    Updated task\n\nRaises:\n    401: If JWT token is invalid or missing\n    404: If task not found or doesn't belong to user","operationId":"toggle_task_completion_api_v1_tasks__task_id__complete_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToggleCompletionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/{user_id}/chat":{"post":{"tags":["Chat"],"summary":"Send Chat Message","description":"Send a chat message and receive AI response.\n\nThe AI agent processes the message, calls MCP tools as needed,\nand returns a friendly response.\n\nArgs:\n    user_id: User ID (must match JWT sub claim)\n    request: Chat request with message\n\nReturns:\n    AI response and conversation ID","operationId":"send_chat_message_api__user_id__chat_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/{user_id}/conversations":{"get":{"tags":["Chat"],"summary":"Get Conversations","description":"Get conversation history for a user.\n\nArgs:\n    user_id: User ID (must match JWT sub claim)\n    limit: Maximum messages to return (default: 50)\n\nReturns:\n    Conversation ID and messages list","operationId":"get_conversations_api__user_id__conversations_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Conversations Api  User Id  Conversations Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AuthResponse":{"properties":{"user":{"$ref":"#/components/schemas/UserResponse"},"token":{"type":"string","title":"Token"}},"type":"object","required":["user","token"],"title":"AuthResponse","description":"Authentication response with user and token."},"ChatRequest":{"properties":{"message":{"type":"string","maxLength":10000,"minLength":1,"title":"Message"}},"type":"object","required":["message"],"title":"ChatRequest","description":"Request body for chat endpoint."},"ChatResponse":{"properties":{"message":{"type":"string","title":"Message"},"conversation_id":{"type":"string","title":"Conversation Id"}},"type":"object","required":["message","conversation_id"],"title":"ChatResponse","description":"Response body for chat endpoint."},"CreateTaskRequest":{"properties":{"title":{"type":"string","maxLength":500,"minLength":1,"title":"Title","description":"Task title"},"description":{"anyOf":[{"type":"string","maxLength":10000},{"type":"null"}],"title":"Description","description":"Task description"},"category":{"anyOf":[{"type":"string","maxLength":50},{"type":"null"}],"title":"Category","description":"Task category","default":"General"},"priority":{"anyOf":[{"type":"string","maxLength":20},{"type":"null"}],"title":"Priority","description":"Task priority","default":"Low"},"due_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Due Date","description":"Due date for the task (ISO format)"}},"type":"object","required":["title"],"title":"CreateTaskRequest","description":"Request model for creating a task."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SigninRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"SigninRequest","description":"Signin request payload."},"SignupRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","minLength":8,"title":"Password","description":"Password must be at least 8 characters"}},"type":"object","required":["email","password"],"title":"SignupRequest","description":"Signup request payload."},"TaskListResponse":{"properties":{"tasks":{"items":{"$ref":"#/components/schemas/TaskResponse"},"type":"array","title":"Tasks"},"count":{"type":"integer","title":"Count"}},"type":"object","required":["tasks","count"],"title":"TaskListResponse","description":"Response model for task list."},"TaskResponse":{"properties":{"id":{"type":"string","title":"Id"},"user_id":{"type":"string","title":"User Id"},"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"completed":{"type":"boolean","title":"Completed"},"category":{"type":"string","title":"Category"},"priority":{"type":"string","title":"Priority"},"due_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Due Date"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","user_id","title","description","completed","category","priority","due_date","created_at","updated_at"],"title":"TaskResponse","description":"Response model for a task."},"ToggleCompletionRequest":{"properties":{"completed":{"type":"boolean","title":"Completed","description":"New completion status"}},"type":"object","required":["completed"],"title":"ToggleCompletionRequest","description":"Request model for toggling task completion."},"TokenResponse":{"properties":{"token":{"type":"string","title":"Token"}},"type":"object","required":["token"],"title":"TokenResponse","description":"Token refresh response."},"UpdateTaskRequest":{"properties":{"title":{"anyOf":[{"type":"string","maxLength":500,"minLength":1},{"type":"null"}],"title":"Title","description":"Task title"},"description":{"anyOf":[{"type":"string","maxLength":10000},{"type":"null"}],"title":"Description","description":"Task description"},"category":{"anyOf":[{"type":"string","maxLength":50},{"type":"null"}],"title":"Category","description":"Task category"},"priority":{"anyOf":[{"type":"string","maxLength":20},{"type":"null"}],"title":"Priority","description":"Task priority"},"due_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Due Date","description":"Due date for the task (ISO format)"}},"type":"object","title":"UpdateTaskRequest","description":"Request model for updating a task."},"UserResponse":{"properties":{"id":{"type":"string","title":"Id"},"email":{"type":"string","title":"Email"},"email_verified":{"type":"boolean","title":"Email Verified"},"created_at":{"type":"string","title":"Created At"},"updated_at":{"type":"string","title":"Updated At"}},"type":"object","required":["id","email","email_verified","created_at","updated_at"],"title":"UserResponse","description":"User response payload."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}